+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -243,6 +243,7 @@ my $expected_tested_by = [
     't/1349-isf-actor-param-package-constants.t',
     't/1350-isf-transaction-param-package-constants.t',
     't/1351-isf-activation-param-package-constants.t',
+    't/1352-isf-library-use-package-constants.t',
 ];
 
 subtest 'direct ISF tested_by metadata is exact and valid' => sub {